requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
Pillow==10.4.0
google-genai==0.3.0
python-multipart==0.0.9
